fix(skills): keep the community tab off the skills loader
The shared shell mounted llmSkillsLogic, so opening Community fetched the skills list plus one count probe per category. The tab definitions and the count probes move to a new skillTabsLogic that loads nothing else, and the shell mounts that.

The tab bar no longer carries tab content either. It used to appear once the counts resolved, which moved the scene content in the tree and remounted it - dropping focus, and cancelling an import whose file picker was still open.

Generated-By: PostHog Desktop
Task-Id: f36ecbdf-860e-40b1-b14c-dbf6e457fe22
