+++ products/skills/frontend/llmSkillsLogic.ts
@@ -28,6 +28,15 @@ import type {
     PaginatedLLMSkillListListApi,
 } from 'products/skills/frontend/generated/api.schemas'
 
+import {
+    DEFAULT_SKILLS_TAB_KEY,
+    SKILL_CATEGORY_TABS,
+    skillCategoryForTabKey,
+    skillTabDescription,
+    skillTabUrl,
+    skillTabsLogic,
+} from './skillTabsLogic'
+
 /** Builds the `/plugin marketplace add` command. The token is embedded once minted; until then
  * a placeholder is shown. */
 export function buildMarketplaceCommand(token: string | null): string {
@@ -70,58 +79,6 @@ export const SKILLS_GROUP_LIMIT = 500
 // as leaves of the deepest allowed group.
 export const SKILLS_GROUP_MAX_DEPTH = 4
 export const LLM_SKILLS_FORCE_RELOAD_PARAM = 'llm_skills_force_reload'
-
-/** URL/UI key for the default tab — the uncategorized skills (everything not pulled into a category tab). */
-export const DEFAULT_SKILLS_TAB_KEY: string = 'skills'
-
-/** Sub-title shown for the default "Skills" tab. Category tabs carry their own copy (below). */
-export const DEFAULT_SKILLS_TAB_DESCRIPTION = 'Manage versioned agent skills that any agent can discover and use.'
-
-export interface SkillCategoryTab {
-    /** Tab key, also the `/skills/<key>` URL segment. */
-    key: string
-    /** The `LLMSkill.category` value this tab filters to. */
-    category: string
-    label: string
-    /** Sub-title shown under the scene title while this tab is active. */
-    description: string
-}
-
-/**
- * Category-backed tabs shown alongside the default "Skills" tab. Each entry surfaces one
- * `LLMSkill.category` value as its own tab and removes it from the default list (which only shows
- * uncategorized skills). To add a tab — e.g. official AI-plugin skills — add an entry here, register
- * the matching `/skills/<key>` route in manifest.tsx, and have the producer stamp that `category`.
- */
-export const SKILL_CATEGORY_TABS: SkillCategoryTab[] = [
-    {
-        key: 'scouts',
-        category: 'scout',
-        label: 'Scouts',
-        description:
-            "Scouts — scheduled agents that scan your project and surface findings in your inbox. Includes PostHog's canonical scouts and any custom ones you author.",
-    },
-    {
-        key: 'review-hog',
-        category: 'review_hog',
-        label: 'Code review',
-        description:
-            "ReviewHog's code-review skills: the specialist perspectives applied in parallel when reviewing a pull request (logic & correctness, contracts & security, performance & reliability) and the validation criteria that decide which findings are worth surfacing. Edit a skill to retune it for your team.",
-    },
-]
-
-export function skillCategoryForTabKey(tabKey: string): string {
-    return SKILL_CATEGORY_TABS.find((tab) => tab.key === tabKey)?.category ?? ''
-}
-
-export function skillTabDescription(tabKey: string): string {
-    return SKILL_CATEGORY_TABS.find((tab) => tab.key === tabKey)?.description ?? DEFAULT_SKILLS_TAB_DESCRIPTION
-}
-
-/** The `/skills` or `/skills/<categoryKey>` path for a tab key. */
-export function skillTabUrl(tabKey: string): string {
-    return tabKey === DEFAULT_SKILLS_TAB_KEY ? urls.skills() : urls.skillsCategoryTab(tabKey)
-}
 
 export interface SkillFilters {
     page: number
@@ -237,8 +194,6 @@ export interface llmSkillsLogicValues {
     activeCategory: string
     activeTabDescription: string
     activeTabKey: string
-    categoryCounts: Record<string, number>
-    categoryCountsLoading: boolean
     codexCommand: string
     connectModalOpen: boolean
     count: number
@@ -255,7 +210,6 @@ export interface llmSkillsLogicValues {
     skills: PaginatedLLMSkillListListApi
     skillsLoading: boolean
     sorting: Sorting | null
-    visibleCategoryTabs: SkillCategoryTab[]
 }
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
@@ -278,31 +232,6 @@ export interface llmSkillsLogicActions {
     }
     issueMarketplaceCommand: (rotate?: boolean) => {
         rotate: boolean
-    }
-    loadCategoryCounts: () => {
-        value: true
-    }
-    loadCategoryCountsFailure: (
-        error: string,
-        errorObject?: any
-    ) => {
-        error: string
-        errorObject?: any
-    }
-    loadCategoryCountsSuccess: (
-        categoryCounts: {
-            [k: string]: number
-        },
-        payload?: {
-            value: true
-        }
-    ) => {
-        categoryCounts: {
-            [k: string]: number
-        }
-        payload?: {
-            value: true
-        }
     }
     loadMarketplaceState: () => {
         value: true
@@ -363,7 +292,6 @@ export interface llmSkillsLogicMeta {
         filters: (rawFilters: Partial<SkillFilters> | null) => SkillFilters
         activeCategory: (activeTabKey: string) => string
         activeTabDescription: (activeTabKey: string) => string
-        visibleCategoryTabs: (categoryCounts: Record<string, number>, activeTabKey: string) => SkillCategoryTab[]
         count: (skills: PaginatedLLMSkillListListApi) => number
         marketplaceCommand: (marketplaceState: LLMSkillMarketplaceCommandApi | null) => string
         codexCommand: (marketplaceState: LLMSkillMarketplaceCommandApi | null) => string
@@ -392,7 +320,6 @@ export const llmSkillsLogic = kea<llmSkillsLogicType>([
             debounce,
         }),
         loadSkills: (debounce: boolean = true) => ({ debounce }),
-        loadCategoryCounts: true,
         setActiveTab: (tabKey: string) => ({ tabKey }),
         deleteSkill: (skillName: string) => ({ skillName }),
         duplicateSkill: (skillName: string, newName: string) => ({ skillName, newName }),
@@ -506,27 +433,6 @@ export const llmSkillsLogic = kea<llmSkillsLogicType>([
                 },
             },
         ],
-        // Per-category skill counts, used to decide which category tabs to show. Loaded once on
-        // mount (and after a delete) with a cheap limit=1 probe per category — we only read `count`.
-        categoryCounts: [
-            {} as Record<string, number>,
-            {
-                loadCategoryCounts: async () => {
-                    const teamId = String(ApiConfig.getCurrentTeamId())
-                    const entries = await Promise.all(
-                        SKILL_CATEGORY_TABS.map(async (tab) => {
-                            const { count } = await llmSkillsList(teamId, {
-                                category: tab.category,
-                                limit: 1,
-                                offset: 0,
-                            })
-                            return [tab.category, count] as const
-                        })
-                    )
-                    return Object.fromEntries(entries)
-                },
-            },
-        ],
     })),
 
     selectors({
@@ -543,17 +449,6 @@ export const llmSkillsLogic = kea<llmSkillsLogicType>([
         activeTabDescription: [
             (s) => [s.activeTabKey],
             (activeTabKey: string): string => skillTabDescription(activeTabKey),
-        ],
-
-        // A category tab is shown only once the team has at least one skill in that category.
-        // The active tab is always kept visible so direct navigation to /skills/<key> (or deleting
-        // the last skill while viewing the tab) doesn't strand the user on a tab that isn't listed.
-        visibleCategoryTabs: [
-            (s) => [s.categoryCounts, s.activeTabKey],
-            (categoryCounts: Record<string, number>, activeTabKey: string): SkillCategoryTab[] =>
-                SKILL_CATEGORY_TABS.filter(
-                    (tab) => (categoryCounts[tab.category] ?? 0) > 0 || tab.key === activeTabKey
-                ),
         ],
 
         count: [(s) => [s.skills], (skills: PaginatedLLMSkillListListApi) => skills.count],
@@ -648,7 +543,7 @@ export const llmSkillsLogic = kea<llmSkillsLogicType>([
                 lemonToast.info(`${skillName || 'Skill'} has been archived.`)
                 await asyncActions.loadSkills(false)
                 // Archiving may have removed the last skill in a category — refresh tab visibility.
-                actions.loadCategoryCounts()
+                skillTabsLogic.findMounted()?.actions.loadCategoryCounts()
             } catch (e) {
                 console.error('Failed to archive skill', e)
                 lemonToast.error('Failed to archive skill')
@@ -784,6 +679,5 @@ export const llmSkillsLogic = kea<llmSkillsLogicType>([
 
     afterMount(({ actions }) => {
         actions.loadSkills()
-        actions.loadCategoryCounts()
     }),
 ])
